# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bills:
    get:
      operationId: getBills
      tags: [bills]
      summary: Get user's bills
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create a new bill
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
...
  /bills/{billId}:
    get:
      operationId: getBill
      tags: [bills]
      summary: Get a bill by ID
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateBill
      tags: [bills]
      summary: Update a bill
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBillRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
    patch:
      operationId: patchBill
      tags: [bills]
      summary: Edit bill header (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchBillRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Only the bill owner can edit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Bill not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBill
      tags: [bills]
      summary: Delete a bill
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
  /bills/{billId}/lines:
    get:
      operationId: getBillLines
      tags: [bill-lines]
      summary: Get bill line items
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillLine"
    post:
      operationId: createBillLine
      tags: [bill-lines]
      summary: Add a line item
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillLineRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillLine"
  /bills/{billId}/lines/bulk:
    post:
      operationId: bulkCreateBillLines
      tags: [bill-lines]
      summary: Bulk create line items (from OCR)
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lines:
                  type: array
                  items:
                    $ref: "#/components/schemas/CreateBillLineRequest"
              required:
                - lines
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillLine"
  /bills/{billId}/lines/{lineId}:
    put:
      operationId: updateBillLine
      tags: [bill-lines]
      summary: Update a line item
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
        - name: lineId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillLineRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillLine"
    delete:
      operationId: deleteBillLine
      tags: [bill-lines]
      summary: Delete a line item
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
        - name: lineId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
  /bills/{billId}/lines/{lineId}/users:
    post:
      operationId: toggleBillLineUser
      tags: [bill-lines]
      summary: Assign or unassign a bill user to a line item
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
        - name: lineId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                billUserId:
                  type: integer
              required:
                - billUserId
      responses:
        "200":
          description: OK (assigned = true, unassigned = false)
          content:
            application/json:
              schema:
                type: object
                properties:
                  assigned:
                    type: boolean
                required:
                  - assigned
  /bills/{billId}/users:
    get:
      operationId: getBillUsers
      tags: [bill-users]
      summary: Get bill participants
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillUser"
    post:
      operationId: createBillUser
      tags: [bill-users]
      summary: Add a participant
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillUser"
  /bills/{billId}/users/{userId}:
    put:
      operationId: updateBillUser
      tags: [bill-users]
      summary: Update a participant (e.g. tipPercentOverride)
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBillUserRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillUser"
    delete:
      operationId: deleteBillUser
      tags: [bill-users]
      summary: Remove a participant
      parameters:
        - name: billId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Bill:
      type: object
      properties:
        id:
          type: integer
        ownerUserId:
          type: integer
          nullable: true
        title:
          type: string
        restaurantName:
          type: string
          nullable: true
        date:
          type: string
        currency:
          type: string
          nullable: true
        taxPercent:
          type: number
        tipPercent:
          type: number
        joinCode:
          type: string
        createdAt:
          type: string
          format: date-time
        settled:
          type: boolean
          description: True when every line item on the bill has been assigned to at least one person
        users:
          type: array
          description: Lightweight participant summary for list rendering (id, name, color)
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              color:
                type: string
            required:
              - id
              - name
              - color
      required:
        - id
        - title
        - date
        - taxPercent
        - tipPercent
        - joinCode
        - createdAt
    BillDetail:
      type: object
      properties:
        bill:
          $ref: "#/components/schemas/Bill"
        lines:
          type: array
          items:
            $ref: "#/components/schemas/BillLine"
        users:
          type: array
          items:
            $ref: "#/components/schemas/BillUser"
        isOwner:
          type: boolean
      required:
        - bill
        - lines
        - users
        - isOwner
    CreateBillRequest:
      type: object
      properties:
        title:
          type: string
        restaurantName:
          type: string
          nullable: true
        date:
          type: string
        currency:
          type: string
          nullable: true
        taxPercent:
          type: number
        tipPercent:
          type: number
      required:
        - title
        - date
        - taxPercent
        - tipPercent
    UpdateBillRequest:
      type: object
      properties:
        title:
          type: string
        restaurantName:
          type: string
          nullable: true
        date:
          type: string
        currency:
          type: string
          nullable: true
        taxPercent:
          type: number
        tipPercent:
          type: number
    PatchBillRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        restaurantName:
          type: string
          nullable: true
        date:
          type: string
        currency:
          type: string
          nullable: true
...
    BillUser:
      type: object
      properties:
        id:
          type: integer
        billId:
          type: integer
        name:
          type: string
        color:
          type: string
        tipPercentOverride:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - billId
        - name
        - color
        - createdAt
    CreateBillUserRequest:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
      required:
        - name
        - color
    UpdateBillUserRequest:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        tipPercentOverride:
          type: number
          nullable: true
    BillLine:
      type: object
      properties:
        id:
          type: integer
        billId:
          type: integer
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
        assignedUserIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - billId
        - description
        - quantity
        - unitPrice
        - total
        - assignedUserIds
        - createdAt
    CreateBillLineRequest:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
      required:
        - description
        - quantity
        - unitPrice
        - total
...